Ajout de la détection de collisions #32

+++ player/src/main/java/fr/unice/polytech/si3/qgl/Mugiwara_Cook/geometry/Point.java
@@ -1,31 +1,31 @@
 package fr.unice.polytech.si3.qgl.Mugiwara_Cook.geometry;
 
 public class Point {
-    int x;
-    int y;
+    double x;
+    double y;
 
     public Point(){
         //Json
         }
 
-    public Point(int x, int y){
+    public Point(double x, double y){
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
